requirements.txt
djangorestframework~=3.15.2
Django~=5.1
drf-nested-routers
django-cors-headers